Add --pinecone-mode= argument
Add an argument --pinecone-mode=rest,sdk,sdk+grpc which can be used to
specify how the test should connect to the Pinecone index.

Re-structure the User classes so all three request types (query,
fetch, delete) are implemented for all connection modes.

Note the default connection mode remains unchanged as 'rest', however
this is expected to change to 'sdk+grpc' (which is faster) in the
near-future...