Remove deprecated 'queries' param from REST mode
The 'queries' parameter to /query is deprecated, and requests using it
fail with a 400 error. Remove it (and use 'vector' instead). Update
integration tests to check that requests are actually successful (so
issues like this can be caught by CI job).

Restructure test_requests.py to remove much of the duplication by
introducing a helper method to actually run the tests.