Fix multiprocess worker heartbeat issues with large dataset loads
When running in multiprocess mode and specifying a --pinecone-dataset,
if the dataset is large (takes >1s to load from disk) then worker
heartbeats can be missed and the master locuust instance can ends up
terminating the workers:

    locust-swarm-0-f008c40/INFO/locust.runners: Worker locust-swarm-0-xxx failed to send heartbeat, setting state to missing.
    locust-swarm-0-f008c40/INFO/locust.runners: Worker locust-swarm-0-yyy failed to send heartbeat, setting state to missing.
    locust-swarm-0-f008c40/INFO/locust.runners: The last worker went missing, stopping test.

This is due to the fact that dataset loading runs in a gevent
greenlet, and the loading into a Pandas DataFrame can block the event
loop for longer than the heartbeat interval (1s). As such, the
heartbeat request message from the master cannot be responded to and
the master has it's timeout expire.

Fix this by running the worker dataset loading in a separate (real)
thread, which does not interfere with the greenlets.

To ensure that Users do not start executing requests until the
dataload is completed, wrap the background thread loading in it's
_own_ greenlet, which the User's __init__ method can wait on. (This
doesn't block the gevent event loop, as the waiting is done at the
greenlet level).