Don't attempt to download directories of datasets
When downloading a Dataset from GCS, skip any blobs which are just
directories. This can cause issues when constructing the parent paths
for actual dataset files, as if try to "download" a directory blob it
is saved as a normal file and hence we cannot later create a
subdirectory of the same name to contain the actual dataset files.