Convert documents to correct schema when dataset has no queries
If a Dataset does not have an explicit 'queries' vectorset, then we
simply re-use the 'documents' vectorset for queries. However, these
two tables have different schemas, and attempting to use 'documents'
for queries directly fails as the 'vector' field does not exist:

 * Documents looks like:
    ["id", "values", "sparse_values", "metadata"]
 * Queries looks like:
    ["vector", "sparse_vector", "filter", "top_k"]

Fix by renaming 'values' to 'vector' when using documents for queries.

Add a testcase to cover this usage, which makes use of
'--pinecone-dataset-ignore-queries' to skip loading the queries set
and hence falls back to documents conversion.