Add --pinecone-dataset-ignore-queries option
Add an option which when specified skips loading the 'queries' set
from a dataset (if present).

This is useful to test datasets which do have a queries set, but the
user would rather make use of the main documents set for queries. It
can also be used to test handling of datasets without queries.