Use sampled 'documents' set for queries when no 'queries' set
If a dataset does not have a 'queries' set included in it (to specify
the vectors to use for querying), then the 'documents' set is used
instead. However, this can be a large volume of data, requiring
significant RAM to hold it (for the entire duration of the test.

This is particularly problematic when running with a large number of
processes as each process must have a copy of the query data.

Address this by adding a new `--pinecone-dataset-docs-sample-for-query`
option which specifies the fraction of documents to use for querying.
The default value is 0.01 (1%).

Example before & after process memory usage (default 0.01 fraction):

    DEBUG/root: Memory usage for pid:1786529 RSS:1491MB VSZ:12351MB pyarrow.allocated:1190MB - On test stop

    DEBUG/root: Memory usage for pid:1787579 RSS:215MB VSZ:12352MB pyarrow.allocated:11MB - On test stop