Pre-convert DataFrame 'values' to Python objects
When loading a Dataset into an index, the 'values' column in the
DataFrame is a pyarrow type (list<item: double>[pyarrow]) as it was
read using the pandas.ArrowDtype types_mapper (see
_load_parquet_dataset).

This _can_ be automatically converted to a Python list object inside
upsert_from_dataframe, but it is slow, as at that level the DataFrame
is iterated row-by-row and the conversion happens one element at a
time.

However, converting the entire sub-frame's column to a Python object
before calling upsert_from_dataframe() is significantly faster, such
that the overall upsert throughput (including the actual server-side
work) is around 3x greater if we pre-convert. Throughput when loading
'wikipedia-simple-text-embedding-ada-002-100K':

Before:
    1096.53 vectors/s

After:
    3681.08 vectors/s