Add --pinecone-dataset-limit option
Add a new option to limit the number documents which should be loaded
from a dataset. This allows a workload to be generated based on a
given dataset but at a reduced document count.

Note: If the dataset includes an explicit 'queries' set then that
queries set is used unchanged, and hence Recall may be significanlty
reduced as the vectors the query expects to find nearby may not
exist. As such, it may be desirable to ignore the query set and
instead randomly sample from the (limited) documents set using the
--pinecone-dataset-ignore-queries option.