Add --pinecone-recall option to report Recall
For Vector Search using approximate nearest neighbour, how close the
returned query result(s) are to the exact nearest neighbour is an important
metric to measure.

Add support for calculating Recall@N for queries, by making use of the
exact top_k present in the 'queries' set. Locust doesn't currently
have way to add additional metrics to the statistics reported, so this
patch reports them instead of latencies; expressing the Recall as a
value between 0 and 100 (locust doens't support fractional latency
values).

Example output:

    Response time percentiles (approximated)
    Type     Name                             50%    66%    75%    80%    90%    95%    98%    99%  99.9% 99.99%   100% # reqs
    --------|---------------------------|--------|------|------|------|------|------|------|------|------|------|------|------
    PC gRPC  Vector (Query only)               98     99     99     99     99    100    100    100    100    100    100  10855
    --------|---------------------------|--------|------|------|------|------|------|------|------|------|------|------|------
	     Aggregated                        99    100    100    100    100    100    100    100    100    100    100  21827

(Yes, the title is misleading, but it is not possible to customise
this in Locust).