Skip loading documents if not needed
The dataset contains a 'documents' set - the actual vectors to index,
and optionally a 'queries' set - the vectors to use for querying.
Depending on how locust has been configured, we may need to load none,
one or both of these sets - they can be large and take time to load,
therefore we want to avoid doing any unnecessary work.

* We must load the 'documents' set if we need to populate the index,
  or if there is no explicit 'queries' set and hence we need to sample
  them from the documents set.

* We must load the 'query' set if it exists (and we haven't been
  instructed to skip loading it).  To simplify the inter-dependencies,
  we load each on-demand when needed.