Migrate deps management from Pip to Poetry
Pip doesn't do a good job of separating the direct and transitive
dependancies of a Python application - there is just a single
'requirements.txt' which lists both.

Given other Pinecone libraries (pinecone-python-client) have migrated
to Poetry, do the same here for improved dependancy management and
consistency of tooling.

Note this changes the installation steps - we now use poetry to
install the dependcies and activate the virtual environment.