Disable destructive requests (delete) by default
Destructive request types (currently delete) are not normally
interesting to include in load-testing - for example deleteById()
returns as soon as the delete has been accepted but not necessarily
when the delete has been applied to the index, so the reported latency
is somewhat meaningless. Furthermore, if any docs are modified
(delete) then on a subsequent run the entire index needs repopulating.

As such, remove delete() from the default set of tasks. If delete (or
any other future destuctive requests) is needed, then it can be
enabled via the new option --pinecone-destructive-tasks.